+++ docs/changelog.md
@@ -7,6 +7,7 @@
 Note: This will be the last update for a little while. The vector and raster formats are tough nuts to crack, so I'm going to come to it with fresh eyes.
 
 - Various bug fixes.
+- Renamed `RGBA` to `Color`.
 - Added the `layer.adjusts` property, a library of tweaks and dials that allows, say, `layer.adjusts.white_balance.tint = -0.4`.
 
 ### 0.0.4 (2020-01-15)